refactor(Signin): remove GetMe component; rename and clean up Signin.tsx
